Organize chains code. Separate 'remote' code.
