+++ truss-chains/truss_chains/deployment/code_gen.py
@@ -40,12 +40,12 @@ from truss.base import truss_config
 from truss.contexts.image_builder import serving_image_builder
 from truss.util import path as truss_path
 
-from truss_chains import definitions, framework, model_skeleton, utils
+from truss_chains import definitions, framework, utils
 
-INDENT = " " * 4
+_INDENT = " " * 4
 _REQUIREMENTS_FILENAME = "pip_requirements.txt"
 _MODEL_FILENAME = "model.py"
-_MODEL_CLS_NAME = model_skeleton.TrussChainletModel.__name__
+_MODEL_CLS_NAME = "TrussChainletModel"
 _TRUSS_GIT = "git+https://github.com/basetenlabs/truss.git"
 _TRUSS_PIP_PATTERN = re.compile(
     r"""
@@ -63,9 +63,15 @@ _TRUSS_PIP_PATTERN = re.compile(
     re.VERBOSE,
 )
 
+_MODEL_SKELETON_FILE = (
+    pathlib.Path(__file__).parent.parent.resolve()
+    / "remote_chainlet"
+    / "model_skeleton.py"
+)
+
 
 def _indent(text: str, num: int = 1) -> str:
-    return textwrap.indent(text, INDENT * num)
+    return textwrap.indent(text, _INDENT * num)
 
 
 def _run_simple_subprocess(cmd: str) -> None:
@@ -312,7 +318,7 @@ def _gen_stub_src(chainlet: definitions.ChainletAPIDescriptor) -> _Source:
                 SplitTextInput(inputs=inputs, extra_arg=extra_arg), SplitTextOutput).root
     ```
     """
-    imports = {"from truss_chains import stub"}
+    imports = {"from truss_chains.remote_chainlet import stub"}
     src_parts: list[str] = []
     input_src = _gen_truss_input_pydantic(chainlet)
     _update_src(input_src, src_parts, imports)
@@ -395,7 +401,7 @@ class _SpecifyChainletTypeAnnotation(libcst.CSTTransformer):
 
 def _gen_load_src(chainlet_descriptor: definitions.ChainletAPIDescriptor) -> _Source:
     """Generates AST for the `load` method of the truss model."""
-    imports = {"from truss_chains import stub", "import logging"}
+    imports = {"from truss_chains.remote_chainlet import stub", "import logging"}
     stub_args = []
     for name, dep in chainlet_descriptor.dependencies.items():
         # `dep.name` is the class name, while `name` is the argument name.
@@ -423,7 +429,10 @@ def _gen_load_src(chainlet_descriptor: definitions.ChainletAPIDescriptor) -> _So
 
 def _gen_predict_src(chainlet_descriptor: definitions.ChainletAPIDescriptor) -> _Source:
     """Generates AST for the `predict` method of the truss model."""
-    imports: set[str] = {"from truss_chains import stub"}
+    imports: set[str] = {
+        "from truss_chains.remote_chainlet import stub",
+        "from truss_chains.remote_chainlet import utils",
+    }
     parts: list[str] = []
     def_str = "async def" if chainlet_descriptor.endpoint.is_async else "def"
     input_model_name = _get_input_model_name(chainlet_descriptor.name)
@@ -444,7 +453,7 @@ def _gen_predict_src(chainlet_descriptor: definitions.ChainletAPIDescriptor) -> 
     # Add error handling context manager:
     parts.append(
         _indent(
-            f"with stub.trace_parent(request), stub.exception_to_http_error("
+            f"with stub.trace_parent(request), utils.exception_to_http_error("
             f'chainlet_name="{chainlet_descriptor.name}"):'
         )
     )
@@ -458,7 +467,7 @@ def _gen_predict_src(chainlet_descriptor: definitions.ChainletAPIDescriptor) -> 
         maybe_await = ""
     run_remote = chainlet_descriptor.endpoint.name
     # See docs of `pydantic_set_field_dict` for why this is needed.
-    args = "**stub.pydantic_set_field_dict(inputs)"
+    args = "**utils.pydantic_set_field_dict(inputs)"
     parts.append(
         _indent(f"result = {maybe_await}self._chainlet.{run_remote}({args})", 2)
     )
@@ -474,9 +483,7 @@ def _gen_predict_src(chainlet_descriptor: definitions.ChainletAPIDescriptor) -> 
 def _gen_truss_chainlet_model(
     chainlet_descriptor: definitions.ChainletAPIDescriptor,
 ) -> _Source:
-    skeleton_tree = libcst.parse_module(
-        pathlib.Path(model_skeleton.__file__).read_text()
-    )
+    skeleton_tree = libcst.parse_module(_MODEL_SKELETON_FILE.read_text())
     imports: set[str] = set(
         libcst.Module(body=[node]).code
         for node in skeleton_tree.body
@@ -489,8 +496,7 @@ def _gen_truss_chainlet_model(
     class_definition: libcst.ClassDef = utils.expect_one(
         node
         for node in skeleton_tree.body
-        if isinstance(node, libcst.ClassDef)
-        and node.name.value == model_skeleton.TrussChainletModel.__name__
+        if isinstance(node, libcst.ClassDef) and node.name.value == _MODEL_CLS_NAME
     )
 
     load_src = _gen_load_src(chainlet_descriptor)
@@ -561,14 +567,32 @@ def _make_requirements(image: definitions.DockerImage) -> list[str]:
         )
     pip_requirements.update(image.pip_requirements)
 
-    has_truss_pypy = any(
-        bool(_TRUSS_PIP_PATTERN.match(req)) for req in pip_requirements
+    truss_pypy = next(
+        (req for req in pip_requirements if _TRUSS_PIP_PATTERN.match(req)), None
     )
-    has_truss_git = any(_TRUSS_GIT in req for req in pip_requirements)
 
-    if not (has_truss_git or has_truss_pypy):
+    truss_git = next((req for req in pip_requirements if _TRUSS_GIT in req), None)
+
+    if truss_git:
+        logging.warning(
+            "The chainlet contains a truss version from github as a pip_requirement:\n"
+            f"\t{truss_git}\n"
+            "This could result in inconsistencies between the deploying client and the "
+            "deployed chainlet. This is not recommended for production chains."
+        )
+    if truss_pypy:
+        logging.warning(
+            "The chainlet contains a pinned truss version as a pip_requirement:\n"
+            f"\t{truss_pypy}\n"
+            "This could result in inconsistencies between the deploying client and the "
+            "deployed chainlet. This is not recommended for production chains. If "
+            "`truss` is not manually added as a requirement, the same version as "
+            "locally installed will be automatically added and ensure compatibility."
+        )
+
+    if not (truss_git or truss_pypy):
         truss_pip = f"truss=={truss.version()}"
-        logging.info(
+        logging.debug(
             f"Truss not found in pip requirements, auto-adding: `{truss_pip}`."
         )
         pip_requirements.add(truss_pip)
