+++ truss/cli/cli.py
@@ -587,7 +587,7 @@ def push_chain(
     # These imports are delayed, to handle pydantic v1 envs gracefully.
     from truss_chains import definitions as chains_def
     from truss_chains import framework
-    from truss_chains import remote as chains_remote
+    from truss_chains.deployment import deployment_client
 
     if watch:
         if publish or promote:
@@ -623,14 +623,14 @@ def push_chain(
             remote=remote,
             environment=environment,
         )
-        service = chains_remote.push(
+        service = deployment_client.push(
             entrypoint_cls, options, progress_bar=progress.Progress
         )
 
     if dryrun:
         return
 
-    assert isinstance(service, chains_remote.BasetenChainService)
+    assert isinstance(service, deployment_client.BasetenChainService)
     curl_snippet = _make_chains_curl_snippet(
         service.run_remote_url, options.environment
     )
@@ -675,7 +675,7 @@ def push_chain(
             console.print(deploy_success_text, style="bold green")
             console.print(f"You can run the chain with:\n{curl_snippet}")
             if watch:  # Note that this command will print a startup message.
-                chains_remote.watch(
+                deployment_client.watch(
                     source,
                     entrypoint,
                     name,
@@ -736,7 +736,7 @@ def watch_chains(
     if a chainlet definition in SOURCE is tagged with `@chains.mark_entrypoint`.
     """
     # These imports are delayed, to handle pydantic v1 envs gracefully.
-    from truss_chains import remote as chains_remote
+    from truss_chains.deployment import deployment_client
 
     if user_env:
         raise ValueError("`user_env` is deprecated, use `environment` instead.")
@@ -744,7 +744,7 @@ def watch_chains(
     if not remote:
         remote = inquire_remote_name(RemoteFactory.get_available_config_names())
 
-    chains_remote.watch(
+    deployment_client.watch(
         source,
         entrypoint,
         name,
